Test: prune old LA based on ENV variable (#1335)
This PR prunes the old `LinearAlgebra` module based on an environment
variable that is set in `.ci/run_tests.jl`. Hopefully, this would mean
that tests that are being run without this environment variable would
run as expected without precompilation issues.

The goal of this PR is to ensure that when julia is being built with the
master branch of `LinearAlgebra`, the tests would work without the need
for pruning (i.e., we want to fix the test failures in
https://github.com/JuliaLang/julia/pull/58242).

This would also mean that to run the tests locally, one would need to
set the `JULIA_PRUNE_OLD_LA` environment variable to `true`.

(cherry picked from commit 87d4c8b08e456eaf88a088945576e30c7660350d)