Prune `LinearAlgebra` module in ambiguity test (#1349)
The ambiguity check was being run in a separate process, but it was not
pruning the `LinearAlgebra` module. This was therefore not testing for
ambiguities in the latest code, but the one in the sysimg.

The PR would be able to replicate the test failure seen in
https://buildkite.com/julialang/julia-master/builds/47424#0196c03a-2230-4d64-b404-e04dbfc76c3f

(cherry picked from commit 0b46d5c8c9b53ff8b0fcbf7e23a8f254ff34fc15)