log for dense diagonal matrix with negative elements (#1352)
Fixes a regression introduced in v1.12, where we have a special branch
for diagonal matrices to improve performance. The issue is that log for
`Diagonal` matrices acts elementwise, which, for real matrices, requires
the diagonal elements to be greater than zero. This PR adds an extra
branch to check this.