Run `__init__()` even in sysimage builds
This change makes the LinearAlgebra initialization code run during the
sysimage build, similar to how it would run when using LinearAlgebra at
top-level in a normal package.

In this configuration, we can't rely on the serializer to isolate
side-effects for us, so we have to manually ensure that any mutations
that we perform here are not accidentally persisted to run-time.

Accordingly, the `OncePerProcess` is here as a reminder that any code
in this init have to think about phasing (and be careful not to leave
behind changes from compile-time that would badly impact run-time).