make eigvals(::Diagonal) accept sortby (#1477)
Solves the `Diagonal` part of #1471 and #1450.

I thought it would be better to start with `Diagonal`, as it was the
only part that required a bit of thinking. I think it also makes sense
in itself, since `eigen` accepted `sortby`, just `eigvals` and `eigvecs`
did not. Even though the docstring said none did.

If this is welcome I'll also do `SymTridiagonal`, and then we can
document that eigenvalues are sorted by default.

---------

Co-authored-by: Daniel Karrasch <daniel.karrasch@posteo.de>