Fix GEMM dispatch for complex-real matmul (#1520)
This should fix #1519. The issue on master is that we have specialized
dispatches for arrays of the same eltype, and the complex-real matmul
ends up in `generic_matmatmul!`. This adds an extra method to ensure
that the complex-real case also reaches BLAS.

```julia
julia> A = ones(ComplexF64, 400, 400); B = ones(size(A)); C = similar(A);

julia> @btime mul!($C, $A, $B);
  57.043 ms (0 allocations: 0 bytes) # master
  1.608 ms (0 allocations: 0 bytes) # this PR
```